+++ CHANGELOG.md
@@ -10,6 +10,8 @@ deploy webhooks to print changelog successfully
 That's all, matcher will stop when detects next line started with `###` match
 
 ### 4.6.0 (15.04.2021 - XX.XX.2021)
+* Fix when doors is not found in legacy versions like 1.8.3, so it caused to throw exception
+
 >>> b1
 * Now players who's in a game can't swap items to off hand
 * Players now can't drag special items in theirs inventory